requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.3
django-allauth==0.51.0
django-gmailapi-backend==0.3.2
gcloud==0.18.3
gmail==0.6.3
google==3.0.0
google-api-core==2.10.2
google-api-python-client==2.66.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.1
googleapis-common-protos==1.56.4
gunicorn==20.1.0
httplib2==0.21.0
idna==3.4
jsonpickle==0.9.6
jwcrypto==1.4.2
oauth2client==4.1.2
oauthlib==3.2.2
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
PyJWT==2.6.0
pyparsing==3.0.9
Pyrebase4==4.5.0
python-jwt==4.0.0
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
wrapt==1.14.1
